backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
anthropic==0.18.1
torch==2.2.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
websockets==12.0
aiofiles==23.2.1
